Title and Decorator for Multiplication Table.
Starting from a Pristine, empty, version of Text Formatting we rewrote
from scratch.

Includes multiple run example for "timing" tests, in development we had
a early output situation.  This is no longer the case.

We have a 'dynamic' decoration, in that we let the user of
multiplication decide what the character is to be used.

We also have a small "Golden Table" to check against for almost simplest
table.